Split search cache api from search result cache api.
